Replaced ExtractRelevantSections
Incorporated the HTML extraction code into our version of the Swami program. The program now uses this material instead of the previous material found in ExtractRelevantSections.py

Note: for the program to run, you need Stanford API extracted into the Swami/src/lib/ directory. The file size is too big to upload to GitHub. See the README in Enchanter/Swami/ for instructions on how to download it.