Replaced RegExps for Lets + Multiline
Replaced the Regular Expression matching in the template generation with a match against all POS patterns for assignment statements - see extractAssignmentAndConditionals() in generateTestTemplates.py.

I also incorporated the first category of multiline rules. The program can check for all lines that follow "If x, then..." POS patterns. There is still work to be done in incorporating them as well as getting other multiline rules accounted for.