Redid Assignment Statements
Redid the assignment statements. The database Swami created was fine if you only ever needed the most recent assignment in the rule list. But when going line by line through a list of rules, the most recent value is not necessarily the correct one, particularly when you're dealing with conditional assignments. I spent a while trying to figure out how to make the database account for the different possible values of a variable, and eventually decided it was probably better to carry over the assignments into the test templates. This way. it might be easier to keep track of whether these variables have the values they need to have according to the specifications.