Fixed Issue with Overwriting Test Files
Fixed an issue where the program would give the same testing file name for all templates with the same function name. This issue was causing files to get overwritten and not appear in results. For example,the test files for array(), array(items), and array(len) were all given the same file name, test-file.js, meaning the first two files were overwritten. The program will now check if a file already exists with the current function name. If so, it will add the function arguments to the name to differentiate it and prevent overwriting.