Fixed Abstract Function Filtering/Fixed Missing Templates
Fixed an error in the abstract function filtering for the IsTestable() component of template generation. Program will no longer generate templates for abstract functions. Also fixed program so that Math functions that were not getting caught by our tool or now being included. These functions were caught by base Swami, but becuase the HTML formatting was a little different for the other functions (used unordered lists), they were not getting extracted as relevant sections.

Finally, when generating templates, Enchanter will create a list of all abstract functions that appear in the relevant sections that are not defined in the abstract_functions.js input file. I am working on detecting if a line of code in a template invokes one of these abstract functions and making sure these undefined functions do not appear in the test code.